apps/windows-edge/requirements-edge.txt
sounddevice
webrtcvad
websockets
aiohttp
python-dotenv
structlog
numpy
